relay/requirements.txt
mysql-connector-python==8.3.0
redis==5.0.3
urllib3>=2.0
python_dateutil>=2.8.2
pydantic>=2.11
typing-extensions>=4.7.1
